update
bench: 2899417

+++ src/nnue/nnue.hpp
@@ -114,16 +114,16 @@ Layer 1
 extern int8_t* l1_weights;
 extern int32_t* l1_bias;
 
-void run_L1(int8_t* input, int32_t* output, int bucket);
+void run_L1(uint8_t* input, int32_t* output, int bucket);
 
 /******
 Layer 2
 *******/
 
-extern int16_t* l2_weights;
+extern int32_t* l2_weights;
 extern int32_t* l2_bias;
 
-int32_t run_L2(int16_t* input, int bucket);
+int32_t run_L2(int32_t* input, int bucket);
 
 void init();
 void cleanup();
